+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## September
 
+* September 8 - Send admins a notification when a `Hired::Form` is created #614 @sarvaiyanidhi
 * September 8 – Update license year and note non-code usage #630
 * September 8 – Remove part-time hiring plan from pricing page #628